+++ src/index.ts
@@ -1,6 +1,3 @@
-export * from './Importer';
-export * from './ImportUtils';
-export * from './parser';
+export * from './import';
+export * from './valuetype';
 export {default} from './phovea';
-export * from './valuetype_idtype';
-export * from './valuetypes';
